+++ src/gf2d/gf2d_actor.h
@@ -159,7 +159,7 @@ Uint32 gf2d_actor_get_framecount(Actor *actor);
  * @param flip (optional) if x or y is non-zero it will flip in that direction
  */
 void gf2d_actor_draw(
-	Actor *actor, float frame, GFC_Vector2D position, GFC_Vector2D *scale, GFC_Vector2D *center, float *rotation,
+	Actor *actor, double frame, GFC_Vector2D position, GFC_Vector2D *scale, GFC_Vector2D *center, double *rotation,
 	GFC_Color *color, GFC_Vector2D *flip
 );
 
@@ -170,6 +170,6 @@ void gf2d_actor_draw(
  * @param frame (optional) if provided, the starting frame will be set here
  * @return NULL on error or not found, the action otherwise
  */
-GFC_Action *gf2d_actor_get_action(Actor *actor, const char *name, float *frame);
+GFC_Action *gf2d_actor_get_action(Actor *actor, const char *name, double *frame);
 
 #endif
